Added fetched data to projects page
